requirements.txt
pycrypto==2.6.1
django==2.1.7
paho-mqtt==1.4.0